fix(44): rename description
