+++ scripts/getContractResponseFromDuplose.ts
@@ -1,6 +1,6 @@
 import { instanceofDuplose, Process, Route, type ContractResponse, type Duplose } from "@duplojs/core";
 import { isStepWithResponse } from "./isStepWithResponse";
-import { duploseIsIgnored } from "./ignore/ignoreThisDuplose";
+import { duploseIsIgnored } from "./ignore/ignoreByTypeCodegenDescription";
 
 export function getContractResponseFromDuplose(duplose: Duplose): ContractResponse[] {
 	if (duploseIsIgnored(duplose)) {
