+++ scripts/ignore/ignoreByTypeCodegenDescription.ts
@@ -1,11 +1,11 @@
 import { Description, type Duplose } from "@duplojs/core";
 
-export class IgnoreThisDuplose extends Description {
+export class IgnoreByTypeCodegenDescription extends Description {
 
 }
 
 export function duploseIsIgnored(duplose: Duplose) {
 	return !!duplose.definiton.descriptions.find(
-		(description) => description instanceof IgnoreThisDuplose,
+		(description) => description instanceof IgnoreByTypeCodegenDescription,
 	);
 }
